server/PredictionService/requirements.txt
django>=4.2.20
djangorestframework
pyyaml
requests
django-cors-headers
python-dotenv==1.0.1
yfinance
pandas
numpy
torch
scikit-learn
tqdm
django-redis==5.4.0
redis==5.0.7
pymongo
mongoengine==0.29.1
kafka-python==2.2.8
confluent-kafka==2.10.0